requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow>=10.0.0
python-dotenv==1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
httpx>=0.25.0
PyJWT>=2.8.0
oss2>=2.18.0